+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add new props `error` on `RadioButton` and `Checkbox`.
+- Feature: Add new `error` props on `RadioButton` and `Checkbox`.
 - Feature: Add new `LocationIcon` React component.
 - Feature: Add new `LocationInput` React component.